requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
aiofiles
PyPDF2
python-docx
pytesseract
Pillow
jinja2
pytest